pyproject.toml
[project]
name = "nutrition"
version = "0.1.0"
description = "A Python package for nutritional analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
nut = "nutrition.cli:main"
